core: More refactoring, code cleanup, docs
